pyproject.toml
# pyproject.toml
# Python project configuration.

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "buildarr-radarr"
version = "0.1.0"
description = "Radarr movie PVR plugin for Buildarr "
authors = ["Callum Dickinson <callum.dickinson.nz@gmail.com>"]
license = "GPL-3.0-or-later"
homepage = "https://buildarr.github.io/plugins/radarr"
repository = "https://github.com/buildarr/buildarr-radarr"
documentation = "https://buildarr.github.io/plugins/radarr"
keywords = [
    "buildarr",
    "radarr",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pydantic",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
readme = "README.md"
packages = [{include = "buildarr_radarr"}]

[tool.poetry.plugins."buildarr.plugins"]
"radarr" = "buildarr_radarr.plugin:RadarrPlugin"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
radarr-py = ">=0.3.1,<0.4.0"
buildarr = ">=0.6.0,<0.7.0"

[tool.poetry.group.dev.dependencies]
black = "23.7.0"
mypy = "1.5.1"
types-requests = "2.31.0.2"
mkdocs = "1.5.2"
mkdocstrings = {extras = ["python"], version = "0.22.0"}
ruff = "0.0.287"

[tool.ruff]
fix = true
format = "grouped"
target-version = "py38"
line-length = 100
select = [
    "A",
    "B",
    "BLE",
    "E",
    "F",
    "G",
    "I",
    "INP",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "PTH",
    "RUF",
    "S",
    "T10",
    "T20",
    "W",
    "YTT",
]
extend-select = ["COM812", "COM818", "UP009"]
extend-ignore = [
    "A003",
    "B023",
    "N805",
    "N806",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "RUF012",
]

[tool.ruff.isort]
lines-between-types = 1
combine-as-imports = true
required-imports = ["from __future__ import annotations"]

[tool.black]
line_length = 100

[tool.mypy]
python_version = "3.8"
pretty = true

[[tool.mypy.overrides]]
module = [
    "radarr.*",
]
ignore_missing_imports = true
